refactor: create_scheduler_log_yaml_file
