+++ 2014_ECRTS_federated_original/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 log = "0.4.14"
+chrono = "0.4"
 env_logger = "0.9.0"
 petgraph = "0.6.3"
 clap = { version = "4.2.4", features = ["derive"] }
