+++ 2014_ECRTS_federated_original/src/main.rs
@@ -1,11 +1,33 @@
+use std::fs;
+
+use chrono::{DateTime, Utc};
 use clap::Parser;
 use lib::dag_creator::*;
 use lib::homogeneous;
-use lib::output_log::*;
 use lib::processor::ProcessorBase;
+use log::{info, warn};
 use outputs_result::*;
 mod federated;
 mod outputs_result;
+
+fn create_yaml_file_core(folder_path: &str, file_name: &str) -> String {
+    if fs::metadata(folder_path).is_err() {
+        let _ = fs::create_dir_all(folder_path);
+        info!("Created folder: {}", folder_path);
+    }
+    let file_path = format!("{}/{}.yaml", folder_path, file_name);
+    if let Err(err) = fs::File::create(&file_path) {
+        warn!("Failed to create file: {}", err);
+    }
+    file_path
+}
+
+fn create_scheduler_log_yaml_file(folder_path: &str, sched_name: &str) -> String {
+    let now: DateTime<Utc> = Utc::now();
+    let date = now.format("%Y-%m-%d-%H-%M-%S").to_string();
+    let file_name = format!("{}-{}-log", date, sched_name);
+    create_yaml_file_core(folder_path, &file_name)
+}
 
 /// Application description and arguments definition using clap crate
 #[derive(Parser)]
