+++ 2020_RTSS_cpc_model_based_algorithm/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 log = "0.4.14"
+chrono = "0.4"
 env_logger = "0.9.0"
 petgraph = "0.6.3"
 lib = { path = "../lib"}
