+++ 2020_RTSS_cpc_model_based_algorithm/src/outputs_result.rs
@@ -1,4 +1,4 @@
-use lib::output_log::append_info_to_yaml;
+use lib::util::append_info_to_yaml;
 use serde_derive::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
@@ -30,16 +30,20 @@ mod tests {
     use crate::prioritization_cpc_model;
 
     use super::*;
+    use chrono::{DateTime, Utc};
     use lib::{
         graph_extension::{GraphExtension, NodeData},
         homogeneous::HomogeneousProcessor,
-        output_log::create_scheduler_log_yaml_file,
         processor::ProcessorBase,
         scheduler::DAGSchedulerBase,
         scheduler_creator::{create_scheduler, SchedulerType},
     };
+    use log::{info, warn};
     use petgraph::Graph;
-    use std::{collections::HashMap, fs::remove_file};
+    use std::{
+        collections::HashMap,
+        fs::{self, remove_file},
+    };
 
     fn create_node(id: i32, key: &str, value: i32) -> NodeData {
         let mut params = HashMap::new();
@@ -71,6 +75,25 @@ mod tests {
         dag.add_edge(n1_0, c2, 1);
 
         dag
+    }
+
+    fn create_yaml_file_core(folder_path: &str, file_name: &str) -> String {
+        if fs::metadata(folder_path).is_err() {
+            let _ = fs::create_dir_all(folder_path);
+            info!("Created folder: {}", folder_path);
+        }
+        let file_path = format!("{}/{}.yaml", folder_path, file_name);
+        if let Err(err) = fs::File::create(&file_path) {
+            warn!("Failed to create file: {}", err);
+        }
+        file_path
+    }
+
+    fn create_scheduler_log_yaml_file(folder_path: &str, sched_name: &str) -> String {
+        let now: DateTime<Utc> = Utc::now();
+        let date = now.format("%Y-%m-%d-%H-%M-%S").to_string();
+        let file_name = format!("{}-{}-log", date, sched_name);
+        create_yaml_file_core(folder_path, &file_name)
     }
 
     #[test]
