+++ lib/src/lib.rs
@@ -4,7 +4,6 @@ pub mod fixed_priority_scheduler;
 pub mod graph_extension;
 pub mod homogeneous;
 pub mod log;
-pub mod output_log;
 pub mod processor;
 pub mod scheduler;
 pub mod scheduler_creator;
