+++ lib/src/scheduler.rs
@@ -1,15 +1,28 @@
-use std::collections::VecDeque;
+use std::{collections::VecDeque, fs};
 
 use crate::{
     core::ProcessResult,
     graph_extension::{GraphExtension, NodeData},
     log::*,
-    output_log::*,
     processor::ProcessorBase,
 };
+use chrono::{DateTime, Utc};
+use log::{info, warn};
 use petgraph::graph::{Graph, NodeIndex};
 
 const DUMMY_EXECUTION_TIME: i32 = 1;
+
+fn create_yaml_file_core(folder_path: &str, file_name: &str) -> String {
+    if fs::metadata(folder_path).is_err() {
+        let _ = fs::create_dir_all(folder_path);
+        info!("Created folder: {}", folder_path);
+    }
+    let file_path = format!("{}/{}.yaml", folder_path, file_name);
+    if let Err(err) = fs::File::create(&file_path) {
+        warn!("Failed to create file: {}", err);
+    }
+    file_path
+}
 
 pub trait DAGSchedulerBase<T>
 where
@@ -132,9 +145,15 @@ where
         }
     }
     fn sort_ready_queue(ready_queue: &mut VecDeque<NodeData>);
+    fn create_scheduler_log_yaml_file(folder_path: &str, sched_name: &str) -> String {
+        let now: DateTime<Utc> = Utc::now();
+        let date = now.format("%Y-%m-%d-%H-%M-%S").to_string();
+        let file_name = format!("{}-{}-log", date, sched_name);
+        create_yaml_file_core(folder_path, &file_name)
+    }
     fn dump_log(&self, dir_path: &str, algorithm_name: &str) -> String {
         let sched_name = format!("{}_{}", algorithm_name, self.get_name());
-        let file_path = create_scheduler_log_yaml_file(dir_path, &sched_name);
+        let file_path = Self::create_scheduler_log_yaml_file(dir_path, &sched_name);
         self.get_log().dump_log_to_yaml(&file_path);
 
         file_path
@@ -147,9 +166,15 @@ pub trait DAGSetSchedulerBase<T: ProcessorBase + Clone> {
     fn get_name(&self) -> String;
     fn get_log(&self) -> DAGSetSchedulerLog;
     fn set_log(&mut self, log: DAGSetSchedulerLog);
+    fn create_scheduler_log_yaml_file(folder_path: &str, sched_name: &str) -> String {
+        let now: DateTime<Utc> = Utc::now();
+        let date = now.format("%Y-%m-%d-%H-%M-%S").to_string();
+        let file_name = format!("{}-{}-log", date, sched_name);
+        create_yaml_file_core(folder_path, &file_name)
+    }
     fn dump_log(&self, dir_path: &str, algorithm_name: &str) -> String {
         let sched_name = format!("{}_{}", algorithm_name, self.get_name());
-        let file_path = create_scheduler_log_yaml_file(dir_path, &sched_name);
+        let file_path = Self::create_scheduler_log_yaml_file(dir_path, &sched_name);
         self.get_log().dump_log_to_yaml(&file_path);
 
         file_path
